refactor: fully decoupling of `clingo` solver
You can use other ASP solvers if you like.

Clingo will be required for a server-side project, but not required for a client-side project.
